==> 05_03NCA_Comp(Kaggle)/01_Get_TeamData.ipynb <==
import pandas as pd
import warnings

warnings.filterwarnings("ignore", message="The frame.append method is deprecated and will be removed from pandas in a future version. Use pandas.concat instead.", category=FutureWarning)

DATA_PATH = 'D:\\0Code\\UNSW\\2024Kaggle\\03_NCA_Comp\\dataset\\'

RSresultW = DATA_PATH+'06_WRegularSeasonDetailedResults.csv'
NCAresultW = DATA_PATH+'07_WNCAATourneyDetailedResults.csv'
RSresultW_df = pd.read_csv(RSresultW)
NCAresultW_df = pd.read_csv(NCAresultW)

RSresultW_df['ScoreDiff'] = RSresultW_df['WScore'] - RSresultW_df['LScore']
RSresultW_df.head(5)

NCAresultW_df['ScoreDiff'] = NCAresultW_df['WScore'] - NCAresultW_df['LScore']
NCAresultW_df.head(5)

Wresult = pd.concat([RSresultW_df, NCAresultW_df]).reset_index(drop=True)

all_teams = pd.unique(pd.concat([Wresult['WTeamID'], Wresult['LTeamID']]))
print(f"The total number of teams is {len(all_teams)}.")

data = Wresult.copy()

years = data['Season'].unique()

new_data = pd.DataFrame(columns=['TeamID', 'Season', 'Score', 'FGM', 'FGA', 'FGM3', 'FGA3', 'FTM', 'FTA', 'OR', 'DR', 'Ast', 'TO', 'Stl', 'Blk', 'ScoreVar', 
                                 'OppScore', 'OppFGM', 'OppFGA', 'OppFGM3', 'OppFGA3', 'OppFTM', 'OppFTA', 'OppStl', 'OppBlk', 'WinPct'])

for year in years:
    # 获取当前年份的数据
    year_data = data[data['Season'] == year]
    
    # 获取当前年份参与比赛的球队ID
    all_teams = pd.unique(pd.concat([year_data['WTeamID'], year_data['LTeamID']]))
    
    # 对每个球队进行处理
    for team_id in all_teams:
        # 处理胜利数据
        win_data = year_data[year_data['WTeamID'] == team_id]
        win_stats = win_data[['WScore', 'WFGM', 'WFGA', 'WFGM3', 'WFGA3', 'WFTM', 'WFTA', 'WOR', 'WDR', 'WAst', 'WTO', 'WStl', 'WBlk']]
        win_stats.columns = ['Score', 'FGM', 'FGA', 'FGM3', 'FGA3', 'FTM', 'FTA', 'OR', 'DR', 'Ast', 'TO', 'Stl', 'Blk']
        
        # 处理失败数据
        loss_data = year_data[year_data['LTeamID'] == team_id]
        loss_stats = loss_data[['LScore', 'LFGM', 'LFGA', 'LFGM3', 'LFGA3', 'LFTM', 'LFTA', 'LStl', 'LBlk']]
        loss_stats.columns = ['Score', 'FGM', 'FGA', 'FGM3', 'FGA3', 'FTM', 'FTA', 'Stl', 'Blk']
        
        # 获取对手数据
        win_opponent_stats = win_data[['LScore', 'LFGM', 'LFGA', 'LFGM3', 'LFGA3', 'LFTM', 'LFTA', 'LStl', 'LBlk']]
        win_opponent_stats.columns = ['OppScore','OppFGM', 'OppFGA', 'OppFGM3', 'OppFGA3', 'OppFTM', 'OppFTA', 'OppStl', 'OppBlk']
        loss_opponent_stats = loss_data[['WScore', 'WFGM', 'WFGA', 'WFGM3', 'WFGA3', 'WFTM', 'WFTA', 'WStl', 'WBlk']]
        loss_opponent_stats.columns = ['OppScore', 'OppFGM', 'OppFGA', 'OppFGM3', 'OppFGA3', 'OppFTM', 'OppFTA', 'OppStl', 'OppBlk']
        
        # 合并胜利和失败数据
        team_stats = pd.concat([win_stats, loss_stats], axis=0)
        opponent_stats = pd.concat([win_opponent_stats, loss_opponent_stats], axis=0)
        
        # 计算平均数
        avg_stats = team_stats.mean(axis=0)
        
        # 计算方差
        var_score = team_stats['Score'].var()
        
        # 计算胜率
        total_games = len(win_data) + len(loss_data)
        win_pct = len(win_data) / total_games if total_games > 0 else 0
        
        # 添加TeamID列、Season列、方差列和胜率列
        avg_stats['TeamID'] = team_id
        avg_stats['Season'] = year
        avg_stats['ScoreVar'] = var_score
        avg_stats['WinPct'] = win_pct
        
        # 添加对手统计数据列
        avg_stats = pd.concat([avg_stats, opponent_stats.mean(axis=0)], axis=0)
        
        # 将计算结果添加到新的数据框中
        new_data = new_data.append(avg_stats, ignore_index=True)

# 重新排列列的顺序
new_data = new_data[['TeamID', 'Season', 'Score', 'WinPct', 'FGM', 'FGA', 'FGM3', 'FGA3', 'FTM', 'FTA', 'OR', 'DR', 'Ast', 'TO', 'Stl', 'Blk', 'ScoreVar', 
                     'OppScore', 'OppFGM', 'OppFGA', 'OppFGM3', 'OppFGA3', 'OppFTM', 'OppFTA', 'OppStl', 'OppBlk']]

# 将 TeamID 列的数据类型转换为与原始数据一致的类型
new_data['TeamID'] = new_data['TeamID'].astype(data['WTeamID'].dtype)
new_data['Season'] = new_data['Season'].astype(data['Season'].dtype)



new_data

new_data.to_csv('001_Wdata.csv', index=False)
