==> 05_03NCA_Comp(Kaggle)/02_Get_Dataset.ipynb <==
import pandas as pd
import numpy as np
import warnings

warnings.filterwarnings("ignore", message="The frame.append method is deprecated and will be removed from pandas in a future version. Use pandas.concat instead.", category=FutureWarning)

DATA_PATH = 'D:\\0Code\\UNSW\\2024Kaggle\\03_NCA_Comp\\dataset\\'

### 1. 首先要打乱比赛结果(csv中分为WTeam和LTeam，没有明确标记结果)

RSresultW = DATA_PATH+'04_WRegularSeasonCompactResults.csv'
RSresultW_df = pd.read_csv(RSresultW)

RSresultW_df = RSresultW_df[RSresultW_df['Season']>=2010]
RSresultW_df = RSresultW_df.drop_duplicates(subset=['Season', 'WTeamID', 'LTeamID'])
RSresultW_df

NCAAresultW = DATA_PATH+'05_WNCAATourneyCompactResults.csv'
NCAAresultW_df = pd.read_csv(NCAAresultW)

NCAAresultW_df = NCAAresultW_df[NCAAresultW_df['Season']>=2010]
NCAAresultW_df = NCAAresultW_df.drop_duplicates(subset=['Season', 'WTeamID', 'LTeamID'])
NCAAresultW_df

# 定义函数以随机交换队伍和分数
def shuffle_teams(row):
    if np.random.rand() > 0.5:
        # 如果随机数大于0.5，交换胜利和失败队伍的ID及其分数
        return pd.Series({'Year':row['Season'], 'Team1': row['LTeamID'], 'Score1': row['LScore'], 
                          'Team2': row['WTeamID'], 'Score2': row['WScore'], 
                          'Result': 1 if row['LScore'] > row['WScore'] else (2 if row['LScore'] < row['WScore'] else 0)})
    else:
        return pd.Series({'Year':row['Season'], 'Team1': row['WTeamID'], 'Score1': row['WScore'], 
                          'Team2': row['LTeamID'], 'Score2': row['LScore'], 
                          'Result': 1 if row['WScore'] > row['LScore'] else (2 if row['WScore'] < row['LScore'] else 0)})


RSresultW_df2 = RSresultW_df.apply(shuffle_teams, axis=1)
RSresultW_df2 = RSresultW_df2.drop(['Score1', 'Score2'], axis=1)
RSresultW_df2['TeamPair'] = RSresultW_df2.apply(lambda row: tuple(sorted([row['Team1'], row['Team2']])), axis=1)
RSresultW_df3 = RSresultW_df2.drop_duplicates(subset=['Year','TeamPair'])
# RSresultW_df3 = RSresultW_df3.drop(columns='TeamPair')
RSresultW_df3

NCAAresultW_df2 = NCAAresultW_df.apply(shuffle_teams, axis=1)
NCAAresultW_df2 = NCAAresultW_df2.drop(['Score1', 'Score2'], axis=1)
NCAAresultW_df2['TeamPair'] = NCAAresultW_df2.apply(lambda row: tuple(sorted([row['Team1'], row['Team2']])), axis=1)
NCAAresultW_df2 = NCAAresultW_df2.drop_duplicates(subset=['Year','TeamPair'])
NCAAresultW_df2

Wresult_df = pd.concat([RSresultW_df3, NCAAresultW_df2], ignore_index=True)
Wresult_df = Wresult_df.drop_duplicates(subset=['Year','TeamPair'],keep='last')
Wresult_df = Wresult_df.drop(columns='TeamPair')
Wresult_df

### 2. 根据Year和TeamID将ID转变为详细数据信息

Wdata = pd.read_csv(DATA_PATH+'001_Wdata.csv')
Wdata.head(5)

Wdata[['TeamID','Season','Score']]

# df1 = RSresultW_df2[RSresultW_df2['Year']==2010]
# df2 = Wdata[Wdata['Season']==2010]

df1 = RSresultW_df2.copy()
df2 = Wdata.copy()

merged_df1 = pd.merge(df1, df2, left_on=['Year', 'Team1'], right_on=['Season', 'TeamID'], how='left')
merged_df1 = merged_df1.rename(columns={
    'Score': 'Score1', 
    'FGM': 'FGM1', 
    'FGA': 'FGA1', 
    'FGM3': 'FGM31', 
    'FGA3': 'FGA31', 
    'FTM': 'FTM1', 
    'FTA': 'FTA1', 
    'OR': 'OR1', 
    'DR': 'DR1', 
    'Ast': 'Ast1', 
    'TO': 'TO1', 
    'Stl': 'Stl1', 
    'Blk': 'Blk1', 
    'ScoreVar': 'ScoreVar1', 
    'OppScore': 'OppScore1', 
    'OppFGM': 'OppFGM1', 
    'OppFGA': 'OppFGA1', 
    'OppFGM3': 'OppFGM31', 
    'OppFGA3': 'OppFGA31', 
    'OppFTM': 'OppFTM1', 
    'OppFTA': 'OppFTA1', 
    'OppStl': 'OppStl1', 
    'OppBlk': 'OppBlk1', 
    'WinPct': 'WinPct1'
})


merged_df1

merged_df2 = pd.merge(df1, df2, left_on=['Year', 'Team2'], right_on=['Season', 'TeamID'], how='left')
merged_df2 = merged_df2.rename(columns={
    'Score': 'Score2', 
    'FGM': 'FGM2', 
    'FGA': 'FGA2', 
    'FGM3': 'FGM32', 
    'FGA3': 'FGA32', 
    'FTM': 'FTM2', 
    'FTA': 'FTA2', 
    'OR': 'OR2', 
    'DR': 'DR2', 
    'Ast': 'Ast2', 
    'TO': 'TO2', 
    'Stl': 'Stl2', 
    'Blk': 'Blk2', 
    'ScoreVar': 'ScoreVar2', 
    'OppScore': 'OppScore2', 
    'OppFGM': 'OppFGM2', 
    'OppFGA': 'OppFGA2', 
    'OppFGM3': 'OppFGM32', 
    'OppFGA3': 'OppFGA32', 
    'OppFTM': 'OppFTM2', 
    'OppFTA': 'OppFTA2', 
    'OppStl': 'OppStl2', 
    'OppBlk': 'OppBlk2', 
    'WinPct': 'WinPct2'
})

merged_df2

# 合并merged_df1和merged_df2
final_df = pd.merge(merged_df1, merged_df2, on=['Year', 'Team1', 'Team2', 'Result'], how='left')

# 选择所需的列
columns_to_keep = [
    'Score1', 'FGM1', 'FGA1', 'FGM31', 'FGA31', 'FTM1', 'FTA1', 'OR1', 'DR1', 
    'Ast1', 'TO1', 'Stl1', 'Blk1', 'ScoreVar1', 'OppScore1', 'OppFGM1', 'OppFGA1', 
    'OppFGM31', 'OppFGA31', 'OppFTM1', 'OppFTA1', 'OppStl1', 'OppBlk1', 'WinPct1',
    'Score2', 'FGM2', 'FGA2', 'FGM32', 'FGA32', 'FTM2', 'FTA2', 'OR2', 'DR2', 
    'Ast2', 'TO2', 'Stl2', 'Blk2', 'ScoreVar2', 'OppScore2', 'OppFGM2', 'OppFGA2', 
    'OppFGM32', 'OppFGA32', 'OppFTM2', 'OppFTA2', 'OppStl2', 'OppBlk2', 'WinPct2', 'Result'
]

final_df = final_df[columns_to_keep]

final_df

final_df2 = final_df.dropna()

final_df2.to_csv('002_Wdataset.csv', index=False)
