==> 05_03NCA_Comp(Kaggle)/03_Model.ipynb <==
import pandas as pd
import numpy as np
from sklearn.model_selection import train_test_split

import warnings

warnings.filterwarnings("ignore", message="The frame.append method is deprecated and will be removed from pandas in a future version. Use pandas.concat instead.", category=FutureWarning)

DATA_PATH = 'D:\\0Code\\UNSW\\2024Kaggle\\03_NCA_Comp\\dataset\\'

Wdataset= DATA_PATH+'002_Wdataset.csv'
W_df = pd.read_csv(Wdataset)

W_df

X = W_df.drop('Result', axis=1)
y = W_df['Result']

X_train, X_test, y_train, y_test = train_test_split(X, y, test_size=0.2, random_state=42)

from sklearn.ensemble import RandomForestClassifier
from sklearn.metrics import classification_report, accuracy_score

# 创建随机森林模型实例
rf = RandomForestClassifier(n_estimators=200, random_state=42)

# 训练模型
rf.fit(X_train, y_train)

# 使用测试集进行预测
y_pred = rf.predict(X_test)

# 打印性能指标
print("Accuracy:", accuracy_score(y_test, y_pred))
print("\nClassification Report:\n", classification_report(y_test, y_pred))


from sklearn.linear_model import LogisticRegression
from sklearn.metrics import classification_report, accuracy_score

# 创建逻辑回归模型实例
lr = LogisticRegression(max_iter=1000, random_state=42)

# 训练模型
lr.fit(X_train, y_train)

# 使用测试集进行预测
y_pred = lr.predict(X_test)

# 打印性能指标
print("Accuracy:", accuracy_score(y_test, y_pred))
print("\nClassification Report:\n", classification_report(y_test, y_pred))


from sklearn.ensemble import GradientBoostingClassifier
from sklearn.metrics import classification_report, accuracy_score

# 创建GBM模型实例
gbm = GradientBoostingClassifier(n_estimators=100, learning_rate=0.1, random_state=42)

# 训练模型
gbm.fit(X_train, y_train)

# 使用测试集进行预测
y_pred = gbm.predict(X_test)

# 打印性能指标
print("Accuracy:", accuracy_score(y_test, y_pred))
print("\nClassification Report:\n", classification_report(y_test, y_pred))


import lightgbm as lgb
from sklearn.model_selection import GridSearchCV
from sklearn.metrics import classification_report, accuracy_score

param_grid = {
    'n_estimators': [50, 100, 150],
    'learning_rate': [0.01, 0.1, 0.5],
    'max_depth': [3, 5, 7]
}

# 创建LightGBM模型实例
lgbm = lgb.LGBMClassifier(random_state=42)

# 创建GridSearchCV实例
grid_search = GridSearchCV(estimator=lgbm, param_grid=param_grid, cv=5, scoring='accuracy', verbose=1, n_jobs=-1)

# 执行网格搜索
grid_search.fit(X_train, y_train)

# 最佳参数
print("Best parameters found: ", grid_search.best_params_)

# 使用最佳参数的模型在测试集上的表现
best_model = grid_search.best_estimator_
y_pred = best_model.predict(X_test)



print("Accuracy on test set: ", accuracy_score(y_test, y_pred))
print("\nClassification Report:\n", classification_report(y_test, y_pred))

# 获取特征重要性
feature_importances = best_model.feature_importances_

# 将特征重要性与特征名称配对并打印
features = X_train.columns
for feature, importance in zip(features, feature_importances):
    print(f"Feature: {feature}, Importance: {importance}")
